Updated automated build script

+++ cookbooks/omnibus-build/recipes/default.rb
@@ -1,5 +1,7 @@
 project_name = 'occi-server'
-work_dir = "/home/vagrant/#{project_name}"
+home_dir = '/home/vagrant'
+work_dir = "#{home_dir}/#{project_name}"
+build_env = { "HOME" => home_dir }
 
 bash 'run_omnibus_build' do
   case node[:platform]
@@ -30,16 +32,23 @@ bash 'run_omnibus_build' do
 
   user 'vagrant'
   group 'omnibus'
+  environment build_env
 
   pkg_dir = pkg_dir ? "#{work_dir}/packages/#{pkg_dir}" : "#{work_dir}/packages"
   cwd work_dir
   timeout 7200
 
   code <<-EOC
-    bundle install --binstubs
-    bin/omnibus build #{project_name}
-    mkdir -p #{pkg_dir}
-    cp #{work_dir}/pkg/* #{pkg_dir}
-    rm -r #{work_dir}/pkg
+    #{home_dir}/.bashrc.d/chruby-default.sh
+    chruby-exec 2.1.2 -- bundle install --binstubs
+    chruby-exec 2.1.2 -- bundle exec omnibus build #{project_name}
+
+    if [ $? -eq 0 ]; then
+      mkdir -p #{pkg_dir}
+      cp #{work_dir}/pkg/* #{pkg_dir}
+      rm -r #{work_dir}/pkg
+    else
+      exit 125
+    fi
   EOC
 end
